backend/requirements.txt
fastapi==0.103.2         # Pydantic v1-compatible
uvicorn[standard]==0.30.0
pydantic==1.10.15
SQLAlchemy==2.0.32
asyncpg==0.29.0
redis==5.0.7
httpx==0.27.0
python-multipart==0.0.9
prometheus-client==0.20.0
alembic==1.13.2
psycopg2-binary==2.9.9  # for Alembic migrations
pytest==8.3.2
groq==0.11.0            # Groq Python SDK (OpenAI-compatible)
google-cloud-vision==3.7.2
Pillow==10.4.0          # Image processing for downscaling and format conversion
python-Levenshtein==0.21.1  # Advanced fuzzy string matching for better book title/author matching
